task2/django-blog-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.7
